Both gates green on 6da0f3a: 'Unit tests and dual-flavor compile' SUCCESS and 'Prod and Staging session runtime' SUCCESS.

Merged with a merge commit rather than a squash on purpose: BrightHarbor's cross-reviews pin individual SHAs (e9fb608, 4eb85c7, 71e03ec, 0e4c6c4, 38c48df, 0063624, f97c8a4, 6da0f3a) and a squash would invalidate every one of those references.

Release is HELD per Kemar — this merges the branch, it does not ship anything.

+++ app/src/main/java/com/ga/airdrop/feature/homedetails/AirCoinViewModel.kt
@@ -19,9 +19,11 @@ data class AirCoinBalanceUiState(
     val status: AirCoinsStatus? = null,
     val loading: Boolean = false,
 ) {
-    val accumulated: Int get() = status?.accumulated ?: status?.balance ?: 0
-    val redeemed: Int get() = status?.redeemed ?: 0
-    val available: Int get() = status?.available ?: status?.balance ?: 0
+    // Doubles: Laravel emits fractional balances (462.25) and the old Int
+    // model truncated them, so a customer silently lost the fraction of a coin.
+    val accumulated: Double get() = status?.accumulated ?: status?.balance ?: 0.0
+    val redeemed: Double get() = status?.redeemed ?: 0.0
+    val available: Double get() = status?.available ?: status?.balance ?: 0.0
 }
 
 class AirCoinBalanceViewModel(
